various cleanup

+++ luther-api.py
@@ -16,9 +16,11 @@ from flask.ext.httpauth import HTTPBasicAuth
 # other imports #
 #################
 
-import dns.query, dns.tsigkeyring, dns.update, dns.exception
+import dns.query, dns.tsigkeyring, dns.update
+from dns.query import UnexpectedSource, BadResponse
 import ipaddress
 import re
+import datetime
 
 ##############################
 # flask + plugin object init #
@@ -33,12 +35,6 @@ db.init_app(app)
 ##################
 # Util functions #
 ##################
-
-def json_status_message(message, code):
-	message = {'status': code, 'message': message}
-	resp = jsonify(message)
-	resp.status_code = code
-	return resp
 
 def validate_ip(ip, v6=False):
 	try:
@@ -57,7 +53,7 @@ def validate_ip(ip, v6=False):
 					correct_subnet = True
 		if test.is_private or not correct_subnet:
 			return False
-		return test.exploeded
+		return test.exploded
 	except ipaddress.AddressValueError:
 		return False
 
@@ -80,57 +76,48 @@ def validate_subdomain(subdomain):
 	else:
 		return False
 
+def json_status_message(message, code, extra=''):
+	if extra is not '':
+		extra = ', '+extra
+	message = {'status': code, 'message': message+extra}
+	resp = jsonify(message)
+	resp.status_code = code
+	return resp
+
 ########################
 # Error handler routes #
 ########################
 
 @app.errorhandler(400)
-def bad_request(error=None, extra=None):
-	if extra:
-		return json_status_message('Bad request, '+extra, 400)
-	else:
-		return json_status_message('Bad request', 400)
+def bad_request(error=None, extra=''):
+	return json_status_message('Bad request', 400, extra)
 
 @app.errorhandler(404)
-def not_found(error=None, extra=None):
-	if extra:
-		return json_status_message('Not found, '+extra, 404)
-	else:
-		return json_status_message('Not found', 404)
+def not_found(error=None, extra=''):
+	return json_status_message('Not found', 404, extra)
 
 @app.errorhandler(405)
-def method_not_allowed(error=None, extra=None):
-	if extra:
-		return json_status_message('Method not allowed, '+extra, 405)
-	else:
-		return json_status_message('Method not allowed', 405)
+def method_not_allowed(error=None, extra=''):
+	return json_status_message('Method not allowed', 405, extra)
 
 @app.errorhandler(409)
-def conflict(error=None, extra=None):
-	if extra:
-		return json_status_message('Conflict in request, '+extra, 409)
-	else:
-		return json_status_message('Conflict in request', 409)
+def conflict(error=None, extra=''):
+	return json_status_message('Conflict in request', 409, extra)
 
-def nothing_to_do(error=None, extra=None):
-	if extra:
-		return json_status_message('Nothing to do, '+extra, 200)
-	else:
-		return json_status_message('Nothing to do', 200)
+def nothing_to_do(error=None, extra=''):
+	return json_status_message('Nothing to do', 200, extra)
 
 @app.errorhandler(204)
-def no_content(error=None, extra=None):
-	if extra:
-		return json_status_message('No content, '+extra, 204)
-	else:
-		return json_status_message('No content', 204)
+def no_content(error=None, extra=''):
+	return json_status_message('No content', 204, extra)
 
 @app.errorhandler(403)
-def forbidden(error=None, extra=None):
-	if extra:
-		return json_status_message('Forbidden, '+extra, 403)
-	else:
-		return json_status_message('Forbidden', 403)
+def forbidden(error=None, extra=''):
+	return json_status_message('Forbidden', 403, extra)
+
+@app.errorhandler(504)
+def upstream_timeout(error=None, extra=''):
+	return json_status_message('Gateway timeout', 504)
 
 #########################
 # DNS setup + functions #
@@ -146,39 +133,54 @@ def update_ddns(name, ip, v6=False):
 		update = dns.update.Update(config.root_domain, keyring=keyring)
 		if not v6:
 			update.replace(name+'.'+config.root_domain, config.default_ttl, 'A', addr)
+			update.replace(name+'.'+config.root_domain, config.default_ttl, 'TXT', '"Record for '+name+'.'+config.root_domain+' last updated at '+str(datetime.datetime.utcnow())+' UTC"')
 		else:
 			update.replace(name+'.'+config.root_domain, config.default_ttl, 'AAAA', addr)
+			update.replace(name+'.'+config.root_domain, config.default_ttl, 'TXT', '"Record for '+name+'.'+config.root_domain+' last updated at '+str(datetime.datetime.utcnow())+' UTC"')
 		try:
-			resp = dns.query.tcp(update, config.dns_master_server, port=config.dns_master_server, source_port=config.dns_master_source_port)
-		except dns.exception.UnexpectedSource or dns.exception.BadResponse:
+			resp = dns.query.tcp(update, config.dns_master_server, port=config.dns_master_port, source_port=config.dns_master_source_port, timeout=config.dns_master_timeout)
+		except UnexpectedSource:
 			return bad_request()
+		except BadResponse:
+			return bad_request()
+		except TimeoutError:
+			return upstream_timeout(extra='tcp connection to master DNS server timed out')
 	else:
 		if not v6:
-			return bad_request('invalid IPv4 address')
+			return 'invalid IPv4 address'
 		else:
-			return bad_request('invalid IPv6 address')
+			return 'invalid IPv6 address'
 
 def new_ddns(name, ip, v6=False):
 	new_record = dns.update.Update(config.root_domain, keyring=keyring)
 	if not v6:
 		new_record.add(name+'.'+config.root_domain, config.default_ttl, 'A', ip)
+		new_record.add(name+'.'+config.root_domain, config.default_ttl, 'TXT', '"Record for '+name+'.'+config.root_domain+' last updated at '+str(datetime.datetime.utcnow())+' UTC"')
 	else:
-		new_record.replace(name+'.'+config.root_domain, config.default_ttl, 'AAAA', ip)
+		new_record.add(name+'.'+config.root_domain, config.default_ttl, 'AAAA', ip)
+		new_record.add(name+'.'+config.root_domain, config.default_ttl, 'TXT', '"Record for '+name+'.'+config.root_domain+' last updated at '+str(datetime.datetime.utcnow())+' UTC"')
 	new_record.absent(name+'.'+config.root_domain)
 	try:
-		resp = dns.query.tcp(new_record, config.dns_master_server, port=config.dns_master_server, source_port=config.dns_master_source_port)
-	except dns.exception.UnexpectedSource or dns.exception.BadResponse:
-		# return 'DNS message error'
+		resp = dns.query.tcp(new_record, config.dns_master_server, port=config.dns_master_port, source_port=config.dns_master_source_port, timeout=config.dns_master_timeout)
+	except UnexpectedSource:
 		return bad_request()
+	except BadResponse:
+		return bad_request()
+	except TimeoutError:
+		return upstream_timeout(extra='tcp connection to master DNS server timed out')
 
 def delete_ddns(name):
 	delete = dns.update.Update(config.root_domain, keyring=keyring)
 	delete.delete(name)
 	delete.present(name)
 	try:
-		resp = dns.query.tcp(delete, config.dns_master_server, port=config.dns_master_server, source_port=config.dns_master_source_port)
-	except dns.exception.UnexpectedSource or dns.exception.BadResponse:
+		resp = dns.query.tcp(delete, config.dns_master_server, port=config.dns_master_port, source_port=config.dns_master_source_port, timeout=config.dns_master_timeout)
+	except UnexpectedSource:
 		return bad_request()
+	except BadResponse:
+		return bad_request()
+	except TimeoutError:
+		return upstream_timeout(extra='tcp connection to master DNS server timed out')
 
 
 ################################
@@ -208,40 +210,50 @@ def get_auth_token():
 
 @app.route('/api/v1/user', methods = ['POST'])
 def new_user():
-    email = request.json.get('email')
-    password = request.json.get('password')
-    if email is None or password is None:
-        return bad_request(extra='missing arguments') # missing arguments
-    if User.query.filter_by(email = email).first() is not None:
-        return conflict(extra='existing user') # existing user
-    user = User(email=email, quota=5, role=1)
-    user.hash_password(password)
-    db.session.add(user)
-    db.session.commit()
-    resp = jsonify({'status': 201, 'email': user.email, 'resources': {'Domains': {'url': 'https://'+config.root_domain+'/api/v1/domain'}, 'Domain updates': 'https://'+config.root_domain+'/api/v1/update'}})
-    resp.status_code = 201
-    return resp
+	if request.json and not request.args:
+	    email = request.json.get('email')
+	    password = request.json.get('password')
+	    if email is None or password is None:
+	        return bad_request('missing arguments') # missing arguments
+	    if User.query.filter_by(email = email).first() is not None:
+	        return conflict('existing user') # existing user
+	    user = User(email=email, quota=5, role=1)
+	    user.hash_password(password)
+	    db.session.add(user)
+	    db.session.commit()
+	    resp = jsonify({'status': 201, 'email': user.email, 'resources': {'Domains': {'url': 'https://'+config.root_domain+'/api/v1/domain'}, 'Domain updates': 'https://'+config.root_domain+'/api/v1/update'}})
+	    resp.status_code = 201
+	    return resp
+	elif request.args and not request.json:
+		pass
+	else:
+		return bad_request('no JSON data or URL Parameters, or both')
 
 @app.route('/api/v1/user', methods = ['DELETE', 'UPDATE'])
 @auth.login_required
 def edit_user():
-	if request.method == 'DELETE':
-		sure = request.json.get('confirm')
-		if sure is None or sure is not 'DELETE':
-			return bad_request(extra='missing or malformed arguments')
-		for d in g.user.domains:
-			delete_ddns(d.name)
-		db.session.delete(g.user)
-		g.user = None
-		db.session.commit()
-		return json_status_message('User deleted, bai bai :<', 200)
-	elif request.method == 'UPDATE':
-		password = request.json.get('password')
-		if password is None:
-			return bad_request(extra='missing arguments')
-		g.user.hash_password(password)
-		db.session.commit()
-		return json_status_message('Password updated', 200)
+	if request.json and not request.args:
+		if request.method == 'DELETE':
+			sure = request.json.get('confirm')
+			if sure is None or sure is not 'DELETE':
+				return bad_request(extra='missing or malformed arguments')
+			for d in g.user.domains:
+				delete_ddns(d.name)
+			db.session.delete(g.user)
+			g.user = None
+			db.session.commit()
+			return json_status_message('User deleted, bai bai :<', 200)
+		elif request.method == 'UPDATE':
+			password = request.json.get('password')
+			if password is None:
+				return bad_request(extra='missing arguments')
+			g.user.hash_password(password)
+			db.session.commit()
+			return json_status_message('Password updated', 200)
+	elif request.args:
+		pass
+	else:
+		return bad_request(extra='no JSON data or URL Parameters, or both')
 
 ################################
 # Domain create / delete route #
@@ -250,59 +262,70 @@ def edit_user():
 @app.route('/api/v1/domains', methods=['GET', 'POST', 'DELETE'])
 @auth.login_required
 def domain_mainuplator():
-	if request.method == 'GET':
-		domains = g.user.domains
-		info = {'email': g.user.email, 'domains': []}
-		for d in domains:
-			domains.append({'domain_name': d.name, 'ip': d.ip, 'domain_token': d.token})
-		if len(info['domains']) > 0:
-			resp = jsonify(info)
-			resp.status_code = 200
-			return resp
-		else:
-			return json_status_message('You have no '+config.root_domain+' subdomains', 200)
-	elif request.method == 'POST':
-		if not g.user.domains.count() == g.user.quota:
-			domain_name = request.json.get('domain_name')
-			if not domain_name:
-				return bad_request('missing arguments')
-			ip = request.json.get('ip')
-			if not ip:
-				ip = request.remote_addr
-			if not validate_subdomain(domain_name):
-				return bad_request('invalid subdomain')
-			if validate_ip(ip):
-				ipv6 = False
+	if request.json and not request.args:
+		if request.method == 'GET':
+			domains = g.user.domains
+			info = {'email': g.user.email, 'domains': []}
+			for d in domains:
+				domains.append({'subdomain': d.name, 'full_domain': d.name+"."+config.root_domain, 'ip': d.ip, 'subdomain_token': d.token})
+			if len(info['domains']) > 0:
+				resp = jsonify(info)
+				resp.status_code = 200
+				return resp
 			else:
-				if validate_ip(ip, v6=True):
-					ipv6 = True
+				return json_status_message('You have no '+config.root_domain+' subdomains', 200)
+		elif request.method == 'POST':
+			if g.user.quota is 0 or not g.user.domains.count() == g.user.quota:
+				domain_name = request.json.get('subdomain')
+				if not domain_name:
+					return bad_request(extra='missing arguments')
+				ip = request.json.get('ip')
+				if not ip:
+					ip = request.remote_addr
+				if not validate_subdomain(domain_name):
+					return bad_request(extra='invalid subdomain')
+				if validate_ip(ip):
+					ipv6 = False
 				else:
-					return bad_request('IP address invalid or not in allowed subnets')
-			new_domain = Domain(name=domain_name, ip=ip, v6=ipv6, user=g.user)
-			new_domain.generate_domain_token()
-			db.session.add(new_domain)
-			db.session.commit()
-			ddns_result = new_ddns(domain_name, ip, ipv6)
-			if not ddns_result:
-				return jsonify({'status': 201, 'subdomain': domain_name, 'full_domain': domain_name+"."+config.root_domain, 'ip': ip, 'domain_token': new_domain.token})
+					if validate_ip(ip, v6=True):
+						ipv6 = True
+					else:
+						return bad_request(extra='IP address invalid or not in allowed subnets')
+				new_domain = Domain(name=domain_name, ip=ip, v6=ipv6, user=g.user)
+				new_domain.generate_domain_token()
+				ddns_result = new_ddns(domain_name, ip, ipv6)
+				if not ddns_result:
+					db.session.add(new_domain)
+					db.session.commit()
+					return jsonify({'status': 201, 'subdomain': domain_name, 'full_domain': domain_name+"."+config.root_domain, 'ip': ip, 'subdomain_token': new_domain.token})
+				else:
+					return ddns_result
 			else:
-				return bad_request(extra=ddns_result)
-		else:
-			return json_status_message('You have reached your subdomain quota', 200)
-	elif request.method == 'DELETE':
-		domain_name = request.json.get('domain_name')
-		domain_token = request.json.get('domain_token')
-		if not domain_name or not domain_token:
-			return bad_request('missing arguments')
-		if not validate_subdomain(domain_name):
-			return bad_request('invalid subdomain')
-		for d in g.user.domains:
-			if d.name is domain_name and d.verify_domain_token(domain_token):
-				db.session.delete(d)
-				db.session.commit()
-				return json_status_message('Subdomain deleted', 200)
-			else:
-				return bad_request('invalid domain_token')
+				return json_status_message('You have reached your subdomain quota', 200)
+		elif request.method == 'DELETE':
+			domain_name = request.json.get('subdomain')
+			domain_token = request.json.get('subdomain_token')
+			if not domain_name or not domain_token:
+				return bad_request(extra='missing arguments')
+			if not validate_subdomain(domain_name):
+				return bad_request(extra='invalid subdomain')
+			for d in g.user.domains:
+				if d.name is domain_name:
+					if d.verify_domain_token(domain_token):
+						ddns_result = delete_ddns(d.name)
+						if not ddns_result:
+							db.session.delete(d)
+							db.session.commit()
+							return json_status_message('Subdomain deleted', 200)
+						else:
+							return ddns_result
+					else:
+						return bad_request(extra='invalid subdomain_token')
+			return bad_request(extra='invalid subdomain')
+	elif request.args and not request.json:
+		pass
+	else:
+		return bad_request(extra='no JSON data or URL Parameters, or both')
 
 
 #################################
@@ -311,7 +334,12 @@ def domain_mainuplator():
 
 @app.route('/api/v1/update', methods=['POST'])
 def fancy_interface():
-	pass
+	if request.json and not request.args:
+		pass
+	elif request.args and not request.json:
+		pass
+	else:
+		return bad_request(extra='no JSON data or URL Parameters, or both')
 
 #########################
 # GET only update route #
@@ -327,23 +355,23 @@ def get_interface(domain_name, domain_token, domain_ip=None):
 			return nothing_to_do(extra='supplied IP is the same as current IP')
 		else:
 			if domain.v6:
-				result = update_ddns(domain, domain_ip, v6=True)
+				ddns_result = update_ddns(domain, domain_ip, v6=True)
 			else:
-				result = update_ddns(domain, domain_ip)
-			if not result:
+				ddns_result = update_ddns(domain, domain_ip)
+			if not ddns_result:
+				domain.ip = domain_ip
+				db.session.commit()
 				return jsonify({'status': 200, 'subdomain': domain_name, 'full_domain': domain_name+"."+config.root_domain, 'ip': domain_ip})
 			else:
-				return result
+				return ddns_result
 	else:
-		return bad_request('invalid domain or token')
+		return bad_request(extra='invalid domain or token')
 
 ##############
 # Dev server #
 ##############
 
 if __name__ == '__main__':
-	app.host = ''
-	app.port = 80
 	app.debug = True
 	with app.app_context():
 		db.create_all()
@@ -351,4 +379,4 @@ if __name__ == '__main__':
 		admin.hash_password(config.default_admin_password)
 		db.session.add(admin)
 		db.session.commit()
-	app.run(use_reloader=False)
+	app.run(use_reloader=False, host='192.168.1.8', port=80)
